moved the PageResponse class to the commons package since its going to be used by other classes too
